requirements.txt
openai==0.28
python-telegram-bot[webhooks]==20.0
